# api/openapi.yaml
openapi: 3.0.3
info:
  title: corps (团队) API
  version: 0.6.0
  description: >
    MVP API for the corps multi-tenant collaboration SaaS. This document is the
    single contract for frontend and backend in Phase 3. All paths are prefixed with
    /api/v1/. Every response uses the uniform envelope { code, data, message }.
    Multi-tenant access control: every {wid} route verifies the caller's membership
    row for that workspace, then runs all tenant queries inside a transaction with
    app.workspace_id / app.user_id set (RLS binds at engine level in hardened
    deployments — see db/schema.sql).
    This contract currently defines 57 paths (updated 2026-09-09).
  contact:
    name: MVP Dev Expert Team — Architecture
servers:
  - url: http://localhost:3000
    description: Local development (Next.js Route Handlers)
  - url: https://ghcr.io/{env}/corps
    description: GHCR Docker deployment
    variables:
      env:
        default: staging
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/workspaces/{wid}/tasks/{id}/decisions:
    get:
      tags: [decisions]
      summary: List decision records for a task
      description: Returns the latest version of each decision linked to the task.
      operationId: listDecisions
      parameters:
        - $ref: '#/components/parameters/Wid'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Decision list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TaskDecisionList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [decisions]
      summary: Create a decision record (version 1)
      description: >
        Creates a decision record (Markdown) linked to the task and appends a version
        history row. Covers AC-10 (version留痕 + bidirectional link).
      operationId: createDecision
      parameters:
        - $ref: '#/components/parameters/Wid'
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDecisionRequest'
      responses:
        '201':
          description: Decision created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Decision'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/workspaces/{wid}/tasks/{id}/decisions/{did}/sync-actions:
    post:
      tags: [decisions]
      summary: Sync decision action items to tasks
      description: >
        Parse action items from decision markdown (checkboxes / "- [ ]" lines)
        and create/update/remove corresponding tasks. Each action item is matched
        by its text hash; new items create tasks, removed items delete stale
        tasks, changed items update task titles. Returns counts of created,
        updated, and removed tasks along with the full task list.
      operationId: syncDecisionActions
      parameters:
        - $ref: '#/components/parameters/Wid'
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/Did'
      responses:
        '200':
          description: Action items synced successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_ActionSyncResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # ANALYTICS & NOTIFICATIONS & EVENTS
  # ===========================================================================
...
components:
  securitySchemes:
    cronSecret:
      type: http
      scheme: bearer
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access JWT (15 min) in Authorization header. workspace_id resolved from JWT claim `wid`.
    refreshAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: >
        Better Auth session cookie (httpOnly, secure, 7-day). The /api/v1/auth/refresh
        endpoint reads this cookie to re-issue a workspace-scoped access JWT.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: HMAC-SHA256 signature of the raw body using the Stripe webhook secret.
  parameters:
    Wid:
      name: wid
      in: path
      required: true
      description: Workspace id. Must equal the JWT `wid` claim; mismatch is 403/404.
      schema: { type: string, format: uuid }
    TaskId:
      name: id
      in: path
      required: true
      description: Task id (UUID).
      schema: { type: string, format: uuid }
...
    Did:
      name: did
      in: path
      required: true
      description: Decision id (UUID).
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Validation error (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 400
            data: null
            message: "email must be a valid address"
    Unauthorized:
      description: Missing or invalid token (401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 401
            data: null
            message: "invalid or expired access token"
    Forbidden:
      description: Authenticated but not permitted (RBAC / tenant mismatch) (403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 403
            data: null
            message: "role 'member' cannot perform this action"
    NotFound:
      description: Resource unknown to caller (RLS-filtered) (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 404
            data: null
            message: "workspace not found"
...
    InternalError:
      description: Internal server error (500)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 500
            data: null
            message: "internal server error"
...
  schemas:
...
    Envelope:
      type: object
      required: [code]
      properties:
        code:
          type: integer
          description: "Business status code, mirrors HTTP status code (200=success, 4xx/5xx=error)"
          example: 200
        data:
          nullable: true
        message:
          type: string
          description: "Human-readable status message. Required for error responses (4xx/5xx), optional for success responses (2xx)."
          example: "ok"
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 401
        data:
          nullable: true
          example: null
        message:
          type: string
          example: "invalid or expired access token"
        # L4：events POST 400 等校验失败响应会附带 zod errors 数组，
        # 此前未在 ErrorEnvelope 中文档化。这里补为可选字段（未列入 required），
        # 既覆盖现有带 errors 的 4xx 响应，也不破坏只返回 {code,message,data} 的错误响应。
        errors:
          type: array
          description: >
            Optional field-level validation errors (e.g. zod issues) for 4xx
            responses. Absent on non-validation errors.
          items:
            type: object
            description: A single validation issue (path/message/code as produced by zod).
...
    Envelope_Decision:
      allOf:
        - $ref: '#/components/schemas/Envelope'
      properties:
        data:
          $ref: '#/components/schemas/Decision'
...
    Envelope_TaskDecisionList:
      allOf:
        - $ref: '#/components/schemas/Envelope'
      properties:
        data:
          type: object
          required: [items, total]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Decision'
            total:
              type: integer
              description: Total count of decision records for this task.
...
    Task:
      type: object
      required: [id, workspaceId, title, status, priority, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          nullable: true
          type: string
          format: uuid
        dueDate:
          nullable: true
          type: string
          format: date-time
        sortOrder:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Decision:
      type: object
      required: [id, taskId, markdown, version, authorId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        markdown:
          type: string
        version:
          type: integer
        authorId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateDecisionRequest:
      type: object
      required: [markdown]
      properties:
        markdown:
          type: string
          minLength: 1
          maxLength: 50000
...
    ActionSyncResponse:
      type: object
      required: [created, updated, removed, tasks]
      properties:
        created:
          type: integer
          description: Number of new tasks created from action items.
          example: 2
        updated:
          type: integer
          description: Number of existing tasks updated (title changed).
          example: 1
        removed:
          type: integer
          description: Number of stale action items removed (tasks deleted).
          example: 0
        tasks:
          type: array
          description: Full list of synced tasks after the operation.
          items:
            $ref: '#/components/schemas/Task'
    Envelope_ActionSyncResponse:
      allOf:
        - $ref: '#/components/schemas/Envelope'
      properties:
        data:
          $ref: '#/components/schemas/ActionSyncResponse'

    # ---- F2: permission matrix ----
...